requirements.txt
discord.py
twilio
